utils/logger.py
import logging
import os
from frontend_bot.config import LOG_LEVEL

LOG_DIR = os.path.join(
    os.path.dirname(os.path.dirname(__file__)), 'logs'
)
os.makedirs(LOG_DIR, exist_ok=True)


def get_logger(name: str) -> logging.Logger:
    logger = logging.getLogger(name)
    if not logger.handlers:
        handler = logging.FileHandler(
            os.path.join(LOG_DIR, f'{name}.log'), encoding='utf-8'
        )
        formatter = logging.Formatter('%(asctime)s %(levelname)s %(message)s')
        handler.setFormatter(formatter)
        logger.addHandler(handler)
        logger.setLevel(getattr(logging, LOG_LEVEL.upper(), logging.INFO))
    return logger 